backend/pyproject.toml
[project]
name = "team-contextor-backend"
version = "0.1.0"
description = "CONTEXTOR local backend (FastAPI) - document upload, parse, multi-agent analysis, persistence."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "sqlalchemy>=2.0.30",
  "aiosqlite>=0.20.0",
  "httpx>=0.27.0",
  "python-docx>=1.1.2",
  "pypdf>=4.2.0",
  "openai>=1.35.0",
  "langgraph>=1.0.5",
  "langchain>=1.2.3",
]
